Move CPU detection code into a separate file
